revert model change for unused fields; refactor tests for consistency while checking attributes
